tests/integration_test.py
import pytest


class TestWalletIntegration:
    """Integration tests for complete wallet flows"""

    @pytest.mark.asyncio
    async def test_complete_deposit_flow(self):
        """Test complete deposit flow: initiate -> webhook -> balance updated"""
        # This would test the full flow in a real database
        pass

    @pytest.mark.asyncio
    async def test_complete_transfer_flow(self):
        """Test complete transfer: check balance -> transfer -> verify balances"""
        pass

    @pytest.mark.asyncio
    async def test_api_key_permissions_enforcement(self):
        """Test API keys with different permissions work correctly"""
        pass


if __name__ == "__main__":
    pytest.main([__file__, "-v"])
