requirements.txt
fastapi[all]
uvicorn
sqlalchemy
asyncpg
pydantic
pydantic-settings
python-dotenv
pyjwt
httpx
authlib
itsdangerous
python-multipart